+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Editing notes always moved them to the "Notes" folder. Now they stay where they started.
 - When you are looking at a particular folder and create a new note, it will be in that folder. (Even the trash!)
+- When opening a document for the first time, the UI would not properly display
 
 ## [0.40.0] - 2020-01-23